move ProxyConfigurationEventBusEvent
