+++ sharding-core/src/main/java/io/shardingsphere/core/orche/eventbus/config/ProxyConfigurationEventBusEvent.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.core.orche.eventbus.config.proxy;
+package io.shardingsphere.core.orche.eventbus.config;
 
 import io.shardingsphere.core.orche.config.ProxyBasicRule;
 import io.shardingsphere.core.rule.DataSourceParameter;
