+++ sharding-jdbc-orchestration/src/main/java/io/shardingsphere/jdbc/orchestration/internal/config/ConfigurationListenerManager.java
@@ -20,7 +20,7 @@ package io.shardingsphere.jdbc.orchestration.internal.config;
 import io.shardingsphere.core.event.ShardingEventBusInstance;
 import io.shardingsphere.core.orche.eventbus.config.MasterSlaveConfigurationEventBusEvent;
 import io.shardingsphere.core.orche.eventbus.config.ShardingConfigurationEventBusEvent;
-import io.shardingsphere.core.orche.eventbus.config.proxy.ProxyConfigurationEventBusEvent;
+import io.shardingsphere.core.orche.eventbus.config.ProxyConfigurationEventBusEvent;
 import io.shardingsphere.core.rule.ShardingRule;
 import io.shardingsphere.jdbc.orchestration.internal.listener.ListenerManager;
 import io.shardingsphere.jdbc.orchestration.internal.state.datasource.DataSourceService;
