+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/config/RuleRegistry.java
@@ -27,7 +27,7 @@ import io.shardingsphere.core.constant.transaction.TransactionType;
 import io.shardingsphere.core.executor.ShardingExecuteEngine;
 import io.shardingsphere.core.metadata.ShardingMetaData;
 import io.shardingsphere.core.orche.config.ProxyBasicRule;
-import io.shardingsphere.core.orche.eventbus.config.proxy.ProxyConfigurationEventBusEvent;
+import io.shardingsphere.core.orche.eventbus.config.ProxyConfigurationEventBusEvent;
 import io.shardingsphere.core.orche.eventbus.state.circuit.CircuitStateEventBusEvent;
 import io.shardingsphere.core.orche.eventbus.state.disabled.DisabledStateEventBusEvent;
 import io.shardingsphere.core.rule.DataSourceParameter;
